Prep for react-router upgrade
